paperspace/requirements.txt
# Paperspace Cloud Dependencies
# Install with: pip install -r requirements.txt

# ============================================================================
# Web Server & API
# ============================================================================
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0
waitress==2.1.2
werkzeug==3.0.1

# ============================================================================
# AI & LLM
# ============================================================================
openai==1.12.0
anthropic==0.18.0
langchain==0.1.7
langchain-openai==0.0.5

# ============================================================================
# Paperspace & Cloud
# ============================================================================
gradient==2.0.6

# ============================================================================
# Google Drive Integration
# ============================================================================
google-api-python-client==2.116.0
google-auth==2.27.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0

# ============================================================================
# Vector Database
# ============================================================================
qdrant-client==1.7.3

# ============================================================================
# Data Processing
# ============================================================================
pandas==2.2.0
numpy==1.26.3
scikit-learn==1.4.0

# ============================================================================
# HTTP & Networking
# ============================================================================
requests==2.31.0
aiohttp==3.9.3
httpx==0.26.0

# ============================================================================
# Document Processing
# ============================================================================
beautifulsoup4==4.12.3
lxml==5.1.0
python-docx==1.1.0
PyPDF2==3.0.1

# ============================================================================
# Utilities
# ============================================================================
python-dotenv==1.0.1
pydantic==2.5.3
python-dateutil==2.8.2
pytz==2024.1

# ============================================================================
# Development & Testing (optional)
# ============================================================================
pytest==7.4.4
pytest-asyncio==0.23.3
black==24.1.1
ruff==0.1.14

